+++ cmake/core-files.cmake
@@ -309,8 +309,6 @@ set(MBGL_CORE_FILES
     include/mbgl/storage/resource_transform.hpp
     include/mbgl/storage/response.hpp
     src/mbgl/storage/asset_file_source.hpp
-    src/mbgl/storage/file_source_request.cpp
-    src/mbgl/storage/file_source_request.hpp
     src/mbgl/storage/http_file_source.hpp
     src/mbgl/storage/local_file_source.hpp
     src/mbgl/storage/network_status.cpp
