+++ platform/qt/config.cmake
@@ -15,7 +15,7 @@ endif()
 
 macro(mbgl_platform_core)
     target_sources(mbgl-core
-        ${MBGL_QT_FILES}
+        ${MBGL_QT_CORE_FILES}
     )
 
     target_include_directories(mbgl-core
@@ -24,7 +24,7 @@ macro(mbgl_platform_core)
     )
 
     target_link_libraries(mbgl-core
-        ${MBGL_QT_LIBRARIES}
+        ${MBGL_QT_CORE_LIBRARIES}
     )
 
     if(NOT WITH_QT_DECODERS)
@@ -50,17 +50,23 @@ macro(mbgl_platform_core)
 
 endmacro()
 
+
+macro(mbgl_filesource)
+    target_sources(mbgl-filesource
+        ${MBGL_QT_FILESOURCE_FILES}
+    )
+
+    target_link_libraries(mbgl-filesource
+        ${MBGL_QT_FILESOURCE_LIBRARIES}
+    )
+endmacro()
+
+
 macro(mbgl_platform_test)
     target_sources(mbgl-test
-        PRIVATE platform/default/mbgl/gl/headless_frontend.cpp
-        PRIVATE platform/default/mbgl/gl/headless_frontend.hpp
-        PRIVATE platform/default/mbgl/gl/headless_backend.cpp
-        PRIVATE platform/default/mbgl/gl/headless_backend.hpp
-        PRIVATE platform/default/mbgl/gl/headless_display.cpp
-        PRIVATE platform/default/mbgl/gl/headless_display.hpp
-        PRIVATE platform/qt/test/headless_backend_qt.cpp
         PRIVATE platform/qt/test/main.cpp
-        PRIVATE platform/qt/test/qmapboxgl.cpp
+        PRIVATE platform/qt/test/qmapboxgl.test.cpp
+        PRIVATE platform/qt/test/qmapboxgl.test.cpp
     )
 
     set_source_files_properties(
@@ -72,7 +78,7 @@ macro(mbgl_platform_test)
 
     target_link_libraries(mbgl-test
         PRIVATE qmapboxgl
-        ${MBGL_QT_TEST_LIBRARIES}
+        PRIVATE mbgl-filesource
     )
 endmacro()
