+++ platform/qt/src/headless_backend_qt.cpp
